+++ intelmq/lib/upgrades.py
@@ -235,6 +235,11 @@ def v202_fixes(defaults, runtime, dry_run):
                 if "query_ripe_stat_ip" not in bot["parameters"]:
                     bot["parameters"]["query_ripe_stat_ip"] = bot["parameters"]["query_ripe_stat_asn"]
                     changed = True
+        if bot["module"] in ("intelmq.bots.experts.cymru_whois.expert",
+                             "intelmq.bots.experts.reverse_dns.expert"):
+            if "overwrite" not in bot["parameters"]:
+                bot["parameters"]["overwrite"] = True
+                changed = True
 
     return changed, defaults, runtime